finish refactor for gif block. relates to #504.
