+++ src/blocks/gif/controls.js
