+++ src/blocks/gif/inspector.js
