+++ src/blocks/gif/size.js
